Refactor `PerfRow` parsing into a reusable module and add Metabase to `docker-compose.yml`
- Moved `PerfRow` struct and parsing logic to `domain::perf` for modularity.
- Updated `import_data_set.rs` to use the new `domain` module for `PerfRow`.
- Introduced Metabase service configuration in `docker-compose.yml` for analytics.
