+++ docker-compose.yml
@@ -11,3 +11,18 @@ services:
       - "5432:5432"
     # volumes:
     #   - ./pg_data:/var/lib/postgresql/data
+
+  metabase:
+    image: metabase/metabase:latest
+    container_name: metabase
+    ports:
+      - "3000:3000"
+    environment:
+      MB_DB_TYPE: postgres
+      MB_DB_DBNAME: ${PG_DB}
+      MB_DB_PORT: 5432
+      MB_DB_USER: ${PG_USER}
+      MB_DB_PASS: ${PG_PASSWORD}
+      MB_DB_HOST: postgres
+    depends_on:
+      - postgres
