+++ src/bin/import_data_set.rs
@@ -3,47 +3,11 @@ use dotenv;
 use pretty_env_logger;
 use csv::ReaderBuilder;
 use sqlx::PgPool;
-use serde::{Deserialize, Deserializer};
+use perf_dashboard::domain::perf::PerfRow;
+
 use anyhow;
 
-fn parse_row_no<'de, D>(deserializer: D) -> Result<Option<i64>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: Option<String> = Option::deserialize(deserializer)?;
-    if let Some(s) = s {
-        let s = s.trim();
-        if s.is_empty() {
-            Ok(None)
-        } else if let Ok(i) = s.parse::<i64>() {
-            Ok(Some(i))
-        } else if let Ok(f) = s.parse::<f64>() {
-            Ok(Some(f as i64))
-        } else {
-            Err(serde::de::Error::custom("Invalid row_no"))
-        }
-    } else {
-        Ok(None)
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct PerfRow {
-    release_tag: String,
-    #[serde(deserialize_with = "parse_row_no")]
-    row_no: Option<i64>,
-    test_scenario: Option<String>,
-    p95_latency_ms: Option<f64>,
-    avg_tps: Option<f64>,
-    peak_tps: Option<f64>,
-    failed_txn_pct: Option<f64>,
-    failed_txn_count: Option<i64>,
-    total_txn_count: Option<f64>,
-    baseline_avg_tps: Option<f64>,
-    test_result_text: Option<String>,
-    remark_text: Option<String>,
-}
-
+/// Import a CSV file into the database.
 pub async fn import_csv(pool: &PgPool, path: &str) -> anyhow::Result<()> {
     let mut rdr = ReaderBuilder::new().from_path(path)?;
     for (i, result) in rdr.deserialize::<PerfRow>().enumerate() {
