+++ src/main.rs
@@ -1,3 +1,5 @@
+mod domain;
+
 fn main() {
     println!("Hello, world!");
 }
